# Improve performance of bottleneck stat functions