<!-- These comments are hidden when you submit the pull request,
so you do not need to remove them! -->

<!-- Please be sure to check out our contributing guidelines,
https://github.com/astropy/astropy/blob/main/CONTRIBUTING.md .
Please be sure to check out our code of conduct,
https://github.com/astropy/astropy/blob/main/CODE_OF_CONDUCT.md . -->

<!-- If you are new or need to be re-acquainted with Astropy
contributing workflow, please see
https://docs.astropy.org/en/latest/development/quickstart.html .
There is even a practical example at
https://docs.astropy.org/en/latest/development/git_edit_workflow_examples.html . -->

<!-- Please just have a quick search on GitHub to see if a similar
pull request has already been posted.
We have old closed pull requests that might provide useful code or ideas
that directly tie in with your pull request. -->

<!-- We have several automatic features that run when a pull request is open.
They can appear daunting but do not worry because maintainers will help
you navigate them, if necessary. -->

### Description
<!-- Provide a general description of what your pull request does.
Complete the following sentence and add relevant details as you see fit. -->

<!-- In addition please ensure that the pull request title is descriptive
and allows maintainers to infer the applicable subpackage(s). -->

<!-- READ THIS FOR MANUAL BACKPORT FROM A MAINTAINER:
Apply "skip-basebranch-check" label **before** you open the PR! -->

The PR improves the performance of bottleneck stat functions when a tuple axis is used by moving the axis keywords to the *last* axis and then combining them so they can be operated on using `axis=-1`.  The current implementation moves the axes to the first axis and then uses `axis=0`, but that operation is slower because the memory is non-contiguous (for the default C-like order).  The performance difference can be significant for larger (or larger ndim) arrays.

Note the tuple axis reordering is necessary because `bottleneck` only allows integer `axis` values.

Simple example show the axis difference:
```python
>>> arr = np.random.rand(500, 500, 500)

>>> %timeit np.nanmedian(a, axis=0)
7.36 s ± 33.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

>>> %timeit np.nanmedian(a, axis=-1)
4.81 s ± 110 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

@mhvk 

<!-- If the pull request closes any open issues you can add this.
If you replace <Issue Number> with a number, GitHub will automatically link it.
If this pull request is unrelated to any issues, please remove
the following line. -->

<!-- Optional opt-out -->

- [ ] By checking this box, the PR author has requested that maintainers do **NOT** use the "Squash and Merge" button. Maintainers should respect this when possible; however, the final decision is at the discretion of the maintainer that merges the PR.