@larrybradley 2024-09-07
The performance difference doesn't seem to be a big for the `bottleneck` `nan*` functions probably due to some internal optimization, but it's not slower (in principle, `axis=-1` is good practice).